backend/tests/requirements.txt
# Testing requirements for Bug Bounty Automation Platform

# Core testing framework
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Additional testing utilities
pytest-xdist>=3.3.1  # Parallel test execution
pytest-html>=3.1.1   # HTML test reports
pytest-sugar>=0.9.7  # Better test output

# Factory and mock utilities
factory-boy>=3.3.0
responses>=0.23.1

# Coverage reporting
coverage>=7.2.0

# Code quality
flake8>=6.0.0
black>=23.7.0

# Performance testing
pytest-benchmark>=4.0.0

# Database testing utilities
pytest-postgresql>=5.0.0  # If needed for PostgreSQL testing

# Security testing
bandit>=1.7.5

# Django testing utilities
django-test-plus>=2.2.0  # Enhanced Django testing
model-bakery>=1.12.0     # Alternative to factory-boy